harvester_node/streamer_world/requirements.txt
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
CouchDB==1.2
decorator==5.1.1
idna==3.4
joblib==1.2.0
Mastodon.py==1.8.1
nltk==3.8.1
python-dateutil==2.8.2
python-magic==0.4.27
regex==2023.5.5
requests==2.28.2
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
